Refactor chunk grid classes to a new module
The chunk grid-related classes were moved from base.core to the newly created chunk_grid module. This change improves code organization, making the codebase easier to maintain and understand. The reference paths in other modules were also updated accordingly.
