+++ src/mdio/schemas/base/core.py
@@ -6,7 +6,6 @@ from typing import get_type_hints
 
 from pydantic import BaseModel
 from pydantic import ConfigDict
-from pydantic import Field
 from pydantic.alias_generators import to_camel
 
 
@@ -14,43 +13,6 @@ class StrictCamelBaseModel(BaseModel):
     """A BaseModel subclass with Pascal Cased aliases."""
 
     model_config = ConfigDict(extra="forbid", alias_generator=to_camel)
-
-
-class RegularChunkShape(StrictCamelBaseModel):
-    """Represents regular chunk sizes along each dimension."""
-
-    chunk_shape: list[int] = Field(
-        ..., description="Lengths of the chunk along each dimension of the array."
-    )
-
-
-class RectilinearChunkShape(StrictCamelBaseModel):
-    """Represents irregular chunk sizes along each dimension."""
-
-    chunk_shape: list[list[int]] = Field(
-        ...,
-        description="Lengths of the chunk along each dimension of the array.",
-    )
-
-
-class RegularChunkGrid(StrictCamelBaseModel):
-    """Represents a rectangular and regularly spaced chunk grid."""
-
-    name: str = Field(default="regular", description="The name of the chunk grid.")
-
-    configuration: RegularChunkShape = Field(
-        ..., description="Configuration of the regular chunk grid."
-    )
-
-
-class RectilinearChunkGrid(StrictCamelBaseModel):
-    """Represents a rectangular and irregularly spaced chunk grid."""
-
-    name: str = Field(default="rectilinear", description="The name of the chunk grid.")
-
-    configuration: RectilinearChunkShape = Field(
-        ..., description="Configuration of the irregular chunk grid."
-    )
 
 
 def model_fields(model: type[BaseModel]) -> dict[str, tuple[Any, Any]]:
